The file_prefix can be also a list

+++ tools/scripts/cli_opentransportdata_csv_fetch_latest.py
@@ -2,6 +2,8 @@ import os, sys
 
 from pathlib import Path
 
+from typing import List
+
 from inc.shared.inc.helpers.config_helpers import load_convenience_config
 from inc.shared.inc.helpers.json_helpers import load_json_from_file
 from inc.shared.inc.helpers.gtfs_helpers import compute_gtfs_day_from_resource_path, compute_gtfs_db_filename
@@ -46,11 +48,12 @@ def _fetch_metadata(script_path, package_id: str):
     print()
     
 def _fetch_resource(script_path: Path, app_config, package_id: str):
-    file_prefix = app_config['csv_latest_data'][package_id].get('file_prefix', None)
-    if file_prefix is None:
+    file_prefix_config = app_config['csv_latest_data'][package_id].get('file_prefix', None)
+    if file_prefix_config is None:
         _fetch_latest_resource(script_path, package_id)
     else:
-        _fetch_resource_by_prefix(app_config, script_path, package_id, file_prefix)
+        file_prefixes = file_prefix_config if isinstance(file_prefix_config, list) else [file_prefix_config]
+        _fetch_resource_by_prefix(app_config, script_path, package_id, file_prefixes)
     #
 
 def _fetch_latest_resource(script_path: Path, package_id: str):
@@ -62,22 +65,24 @@ def _fetch_latest_resource(script_path: Path, package_id: str):
     print(ckan_fetch_sh, flush=True)
     os.system(ckan_fetch_sh)
     
-def _fetch_resource_by_prefix(app_config: any, script_path: Path, package_id: str, resource_prefix: str):
-    ckan_resource = compute_ckan_resource_by_prefix(app_config, package_id, resource_prefix)
+def _fetch_resource_by_prefix(app_config: any, script_path: Path, package_id: str, resource_prefixes: List[str]):
+    for resource_prefix in resource_prefixes:
+        ckan_resource = compute_ckan_resource_by_prefix(app_config, package_id, resource_prefix)
     
-    # fetch latest archive
-    ckan_fetch_cli_path = f'{script_path.parent}/../ckan-utils/fetch_package_cli.py'
-    ckan_fetch_sh = f'{PYTHON_PATH} {ckan_fetch_cli_path} --package_id {package_id} --resource_title {ckan_resource.identifier}'
+        # fetch latest archive
+        ckan_fetch_cli_path = f'{script_path.parent}/../ckan-utils/fetch_package_cli.py'
+        ckan_fetch_sh = f'{PYTHON_PATH} {ckan_fetch_cli_path} --package_id {package_id} --resource_title {ckan_resource.identifier}'
     
-    print(f'STEP {package_id}.2 - FETCH RESOURCE by PREFIX {ckan_resource.identifier}')
-    print(ckan_fetch_sh, flush=True)
-    os.system(ckan_fetch_sh)
+        print(f'STEP {package_id}.2 - FETCH RESOURCE by PREFIX {ckan_resource.identifier}')
+        print(ckan_fetch_sh, flush=True)
+        os.system(ckan_fetch_sh)
+    #
 
     print()
 
 def _symlink_latest_resource(app_config, package_id: str):
-    symlink_latest_path_s: str = app_config['csv_latest_data'][package_id].get('symlink_latest_path', None)
-    if symlink_latest_path_s is None:
+    has_no_symlink = app_config['csv_latest_data'][package_id].get('symlink_latest_path', None) is None
+    if has_no_symlink:
         return
     
     print(f'STEP {package_id}.3 - SYMLINK LATEST DATASET')
@@ -85,31 +90,39 @@ def _symlink_latest_resource(app_config, package_id: str):
     package_base_path: str = app_config['data_paths']['opentransportdata']['package_base_path']
     package_base_path = package_base_path.replace('[PACKAGE_ID]', package_id)
     
-    resource_prefix = app_config['csv_latest_data'][package_id]['file_prefix'] or None
-    ds_resource = compute_ckan_resource_by_prefix(app_config, package_id, resource_prefix)
+    file_prefix_config = app_config['csv_latest_data'][package_id]['file_prefix'] or None
+    if file_prefix_config is None:
+        return
     
-    ds_res_filename = ds_resource.url.split('/')[-1]
-    ds_res_extension = Path(ds_res_filename.lower()).suffix
-    
-    ds_res_path = Path(f'{package_base_path}/{ds_res_filename}')
-    if ds_res_extension == '.zip':
-        ds_res_zip_folder = ds_res_filename[0:-4]
-        ds_res_path = Path(f'{package_base_path}/{ds_res_zip_folder}/{ds_res_zip_folder}')
-    # else:
-    
-    symlink_latest_path_s = symlink_latest_path_s.replace('[PACKAGE_ID]', package_id)
-    file_prefix = app_config['csv_latest_data'][package_id]['file_prefix']
-    symlink_latest_path_s = symlink_latest_path_s.replace('[PREFIX]', file_prefix)
-    
-    symlink_latest_path = Path(symlink_latest_path_s)
-    
-    if os.path.islink(symlink_latest_path):
-        os.remove(symlink_latest_path)
-    os.symlink(ds_res_path, symlink_latest_path)
+    file_prefixes = file_prefix_config if isinstance(file_prefix_config, list) else [file_prefix_config]
+    for file_prefix in file_prefixes:
+        ds_resource = compute_ckan_resource_by_prefix(app_config, package_id, file_prefix)
+        
+        ds_res_filename = ds_resource.url.split('/')[-1]
+        ds_res_extension = Path(ds_res_filename.lower()).suffix
+        
+        ds_res_path = Path(f'{package_base_path}/{ds_res_filename}')
+        if ds_res_extension == '.zip':
+            ds_res_zip_folder = ds_res_filename[0:-4]
+            ds_res_path = Path(f'{package_base_path}/{ds_res_zip_folder}/{ds_res_zip_folder}')
+        #
+        
+        symlink_latest_path_s = app_config['csv_latest_data'][package_id]['symlink_latest_path']
+        symlink_latest_path_s = symlink_latest_path_s.replace('[PACKAGE_ID]', package_id)
+        symlink_latest_path_s = symlink_latest_path_s.replace('[PREFIX]', file_prefix)
+        
+        symlink_latest_path = Path(symlink_latest_path_s)
+        
+        if os.path.islink(symlink_latest_path):
+            os.remove(symlink_latest_path)
+        os.symlink(ds_res_path, symlink_latest_path)
 
+        print()
+        print(f'=> {symlink_latest_path}')
+    # loop prefix
+    
     print()
-    print(f'=> {symlink_latest_path}')
-    print()
+# _symlink_latest_resource
 
 if __name__ == "__main__":
     main()
